external/requirements.txt
# for external reciever.py
requests
urllib3
pydbus
PyGObject
filetype
systemd
python-pidfile